+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/from_request_bodies/mod.rs
@@ -20,14 +20,6 @@ struct Shaper {
 
 impl Shaper {
     fn run(self) -> Result<DefinitionShape> {
-        // DefinitionShape::Enum {
-        //     header: TypeHeaderShape {
-        //         name: self.name,
-        //         doc_comments: (),
-        //         is_nullable: false,
-        //     },
-        //     values: (),
-        // };
         unimplemented!()
     }
 }
