+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/from_schemas/to_type_shape.rs
@@ -1,4 +1,4 @@
-use crate::conversions::v3_0::to_rust_type::TypeShape;
+use super::TypeShape;
 use crate::conversions::Error::UnknownFormat;
 use crate::conversions::Result;
 use crate::targets::rust_type::DataType;
