+++ libs/gesha-core/src/conversions/v3_0/to_rust_type/post_processor/resolve_all_of.rs
@@ -1,5 +1,7 @@
+use crate::conversions::v3_0::to_rust_type::from_schemas::{
+    AllOfItemShape, DefinitionShape, FieldShape,
+};
 use crate::conversions::v3_0::to_rust_type::post_processor::PostProcessor;
-use crate::conversions::v3_0::to_rust_type::{AllOfItemShape, DefinitionShape, FieldShape};
 use crate::conversions::Result;
 
 impl PostProcessor {
